pyproject.toml
[tool.poetry]
name = "mnistops"
version = "0.1.0"
description = "Learning project for MLOps course"
authors = ["Petryashin Ilia <petryashin.i.e@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
scikit-learn = "^1.3.1"
pandas = "^2.0.3"
pathlib = "^1.0.1"
tqdm = "^4.66.1"
torch = "2.0.0"
torchvision = "0.15.1"
dvc = {version = "3.26.2", extras = ["gdrive"]}
hydra-core = "^1.3.2"
pytorch-lightning = "^2.1.0"
mlflow = "^2.7.1"
wandb = "^0.15.12"
onnx = "1.15.0"
onnxruntime = "1.16.3"
tritonclient = {version = "^2.40.0", extras = ["http"]}
fire = "^0.5.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
Flake8-pyproject = "^1.2.3"
ipykernel = "^6.25.2"

[tool.black]
line-length = 80

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 80

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
